[frontend] light refactoring on custom overview settings
